fix: read from existing file and remove reader in dbfile

+++ src/main/java/cn/edu/thssdb/storage/MetaFile.java
@@ -1,5 +1,6 @@
 package cn.edu.thssdb.storage;
 
+import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.FileOutputStream;
@@ -8,6 +9,7 @@ import java.io.ObjectInputStream;
 
 public class MetaFile {
     private String filename;
+    private Metadata data; // newest content of file
     private FileInputStream in;
     private FileOutputStream out;
 
@@ -15,11 +17,18 @@ public class MetaFile {
         this.filename = filename + ".meta";
     }
 
-    public void createFile() {
+    public void init() {
         try {
-            out = new FileOutputStream(this.filename);
-            in = new FileInputStream(this.filename);
-        } catch (IOException e) {
+            if (!(new File(this.filename).isFile())) {
+                out = new FileOutputStream(this.filename);
+                in = new FileInputStream(this.filename);
+            } else {
+                // Could not instanitiate fileoutputstream as it will create a new empty file
+                in = new FileInputStream(this.filename);
+                ObjectInputStream objin = new ObjectInputStream(this.in);
+                this.data = (Metadata) objin.readObject();
+            }
+        } catch (Exception e) {
             System.err.println(String.format("Create file error:%s", this.filename));
             System.exit(-1);
         }
@@ -27,6 +36,10 @@ public class MetaFile {
 
     public void writeMetadata(Metadata data) {
         try {
+            this.data = data;
+            if (this.out == null) {
+                this.out = new FileOutputStream(this.filename);
+            }
             ObjectOutputStream objout = new ObjectOutputStream(this.out);
             objout.writeObject(data);
         } catch (IOException e) {
@@ -35,12 +48,14 @@ public class MetaFile {
     }
 
     public Metadata readMetadata() {
-        Metadata result = null;
-        try {
-            ObjectInputStream objin = new ObjectInputStream(this.in);
-            result = (Metadata) objin.readObject();
-        } catch (Exception e) {
-            System.err.println(String.format("Serialize metadata failed:%s", this.filename));
+        Metadata result = this.data;
+        if (result == null) {
+            try {
+                ObjectInputStream objin = new ObjectInputStream(this.in);
+                result = (Metadata) objin.readObject();
+            } catch (Exception e) {
+                System.err.println(String.format("DeSerialize metadata failed:%s", this.filename));
+            }
         }
         return result;
     }
